First fully completed draft.
